+++ Forum3DView/Forum3DView.cpp
@@ -20,7 +20,7 @@
 
 #ifdef SCAD11
 #include "Result.hpp"
-#include "Grand.h"
+//#include "Grand.h"
 #endif
 #ifdef SCAD21
 //#include "../../Scad11/include/RESULT.HPP"
@@ -296,7 +296,7 @@ extern "C" BOOL __stdcall SCForumMdl3DDraw( HDC hDC, const RECT * r, LPCTSTR sFi
 
 	int cx = (Region.left + Region.right) / 2, cy = (Region.top + Region.bottom) / 2;
 	size_t NumElements = fvg.ElementArray.size();
-	for(int i = 0; i < NumElements /*&& !Terminated*/; i++)
+	for(size_t i = 0; i < NumElements /*&& !Terminated*/; i++)
 	{
 		//int Percent = (k) * 100.0 / nFiles + (i + 1) * 100.0 / (2 * nFiles * NumElements);
